Refactored the player class. Should now have vectors which controls its position.

+++ core/src/com/tda367/game/Interfaces/IObservers.java
@@ -2,5 +2,5 @@ package Interfaces;
 
 public interface IObservers {
 
-    void keyPressed(int key);
+    void keyPressed(float key);
 }
